semana10/dia6/deploy_aws/requirements.txt
django
Pillow
django-admin-tailwind
django-crispy-forms
djangorestframework
python-dotenv
cloudinary
django-paypal
gunicorn
psycopg2
dj-database-url